python/requirements.txt
pytorch
monai
einops
onnx
onnxscript
